+++ examples/triangle/triangle_app.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <cubey/app_config.h>
+
+namespace cubey::examples::triangle {
+
+int run_triangle(const RunConfig& config);
+
+} // namespace cubey::examples::triangle